requirements.txt
torch==1.9.0
pandas==1.3.3
torchaudio==0.9.0a0+33b2469
pytorch_lightning==1.4.6
numpy==1.21.2
tqdm==4.62.2
matplotlib==3.4.3
librosa==0.8.1
pytest==6.2.5
PyYAML==5.4.1
seaborn==0.11.2
